moved plotly plotting utilities into their own module.
